cli/requirements.txt
PyInquirer==1.0.3
PyYAML==5.4.1